0.8.7

+++ tibanna/core.py
@@ -80,6 +80,7 @@ class API(object):
     default_stepfunction_name = TIBANNA_DEFAULT_STEP_FUNCTION_NAME
     default_env = ''
     sfn_type = 'unicorn'
+    do_not_delete = []  # list of lambda names that should not be deleted before updating
 
     def __init__(self):
         pass
@@ -165,7 +166,7 @@ class API(object):
         except Exception as e:
             raise(e)
         # adding execution info to dynamoDB for fast search by awsem job id
-        self.add_to_dydb(jobid, run_name, sfn, data['config']['log_bucket'])
+        self.add_to_dydb(jobid, run_name, sfn, data['config']['log_bucket'], verbose=verbose)
         data[_tibanna]['response'] = response
         if verbose:
             # print some info
@@ -186,7 +187,7 @@ class API(object):
         dydb = boto3.client('dynamodb', region_name=AWS_REGION)
         try:
             # first check the table exists
-            dydb.describe_table(TableName=DYNAMODB_TABLE)
+            res = dydb.describe_table(TableName=DYNAMODB_TABLE)
         except Exception as e:
             if verbose:
                 printlog("Not adding to dynamo table: %s" % e)
@@ -339,7 +340,8 @@ class API(object):
                             res_s3 = boto3.client('s3').get_object(Bucket=logbucket, Key=job_id + suffix)
                         except Exception as e:
                             if 'NoSuchKey' in str(e):
-                                printlog("log/postrunjson file is not ready yet. Wait a few seconds/minutes and try again.")
+                                printlog("log/postrunjson file is not ready yet." +
+                                         "Wait a few seconds/minutes and try again.")
                                 return ''
                             else:
                                 raise e
@@ -622,13 +624,14 @@ class API(object):
             full_function_name = name + '_' + function_name_suffix
         else:
             full_function_name = name
-        try:
-            boto3.client('lambda').get_function(FunctionName=full_function_name)
-            print("deleting existing lambda")
-            boto3.client('lambda').delete_function(FunctionName=full_function_name)
-        except Exception as e:
-            if 'Function not found' in str(e):
-                pass
+        if name not in self.do_not_delete:
+            try:
+                boto3.client('lambda').get_function(FunctionName=full_function_name)
+                print("deleting existing lambda")
+                boto3.client('lambda').delete_function(FunctionName=full_function_name)
+            except Exception as e:
+                if 'Function not found' in str(e):
+                    pass
         aws_lambda.deploy_function(lambda_fxn_module,
                                    function_name_suffix=function_name_suffix,
                                    package_objects=self.tibanna_packages,
